Build ts-angular-components style buttons with react
The thing that really attracted me about react was the fact that it seems to be a lot easier to compose functions to build components. That would help me to build tools like I built for renovo while avoiding some of the pitfalls of excessive inheritance that I fell into there.