Example of using a template provided by the consumer to stamp out data items
My initial reaction was that angular would be superior in this area because react children are fairly limited. Now I'm not so sure. The ability of react to pass around simple functions that turn data into jsx is fantastic. Furthermore, the fact that this all gets strongly typed means that consumers will be clearly told how they need to provide a template. Also, the fact that these can be parameters instead of view children with selectors makes it more explicit as well.