Added a helper for stoping event propagation
The beauty is that with the way react works, we can just make this a function and attach it as a click handler (or any other event). For ease of use, I decided to make it a 'curry' function, so you have to call it, but can optionally call in a function to run 'onClick' after stopping propagation.